Create attribute group manager and move some code from attribute manager to it
